Moved hooksm to utils
